User shouldn't have to provide protocol and/or slash at the end of server url #6636
After this patch user should just be able to provide server URL without
having to specify protocol (http/https), also if there's no '/' at the
end of the URL it'll be appended.

If no protocol is provided, Angular tryton would first try to connect
using https, if it fails it'll connect using http.